backend/functions/search-service/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.13.1
numpy==1.24.3
